requirements.txt
fastapi
sqlmodel
python-dotenv
